pyproject.toml
[project]
name = "langchain-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai>=0.8.5",
    "huggingface-hub>=1.1.6",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.22",
    "langchain-core>=0.3.80",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.35",
    "numpy>=2.0.2",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.6.1",
    "streamlit>=1.50.0",
    "transformers>=2.3.0",
]
